pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "faux-splunk-cloud"
version = "0.1.0"
description = "Ephemeral Splunk Cloud Victoria instances for development and testing"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Faux Splunk Cloud Contributors" }
]
keywords = ["splunk", "cloud", "victoria", "ephemeral", "testing", "development"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    # Web framework for ACS API simulation
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Official Splunk SDK for API compatibility
    "splunk-sdk>=2.0.0",

    # Container orchestration
    "docker>=7.0.0",

    # HTTP client for health checks and API calls
    "httpx>=0.26.0",

    # JWT authentication (ACS API uses JWT)
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",

    # SAML authentication (Keycloak integration)
    "python3-saml>=1.16.0",

    # Configuration and templating
    "pyyaml>=6.0.1",
    "jinja2>=3.1.0",

    # CLI interface
    "typer>=0.9.0",
    "rich>=13.7.0",

    # State persistence
    "aiosqlite>=0.19.0",
    "sqlalchemy[asyncio]>=2.0.0",

    # OpenAPI spec generation (for ACS compatibility)
    "openapi-spec-validator>=0.7.0",

    # Fake data generation for attack simulation
    "faker>=22.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing core
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "pytest-randomly>=3.15.0",

    # BDD / Gherkin
    "pytest-bdd>=7.0.0",

    # Test isolation & mocking
    "testcontainers>=3.7.0",
    "freezegun>=1.4.0",
    "respx>=0.20.0",

    # Test data generation
    "factory-boy>=3.3.0",
    "faker>=22.0.0",

    # Code quality
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "pre-commit>=3.6.0",

    # HTTP testing
    "httpx>=0.26.0",
]

[project.scripts]
faux-splunk = "faux_splunk_cloud.cli:app"
fsc = "faux_splunk_cloud.cli:app"

[project.urls]
Homepage = "https://github.com/daijoubussy/faux-splunk-cloud"
Repository = "https://github.com/daijoubussy/faux-splunk-cloud"

[tool.hatch.build.targets.wheel]
packages = ["src/faux_splunk_cloud"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # Pyflakes
    "I",     # isort
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
    "ARG",   # flake8-unused-arguments
    "SIM",   # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/faux_splunk_cloud --cov-report=term-missing"
markers = [
    "unit: Fast isolated unit tests",
    "integration: Tests requiring external services",
    "e2e: End-to-end workflow tests",
    "slow: Tests taking >5 seconds",
    "docker: Tests requiring Docker daemon",
    "critical: Must pass for deployment",
    "wip: Work in progress, may be skipped",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
timeout = 30
